Finish some query

+++ src/main/java/com/shinonometn/re/ssim/controller/ManagementController.kt
@@ -1,29 +1,52 @@
 package com.shinonometn.re.ssim.controller
 
+import com.shinonometn.re.ssim.commons.CacheKeys
 import com.shinonometn.re.ssim.models.CaptureTaskDTO
 import com.shinonometn.re.ssim.services.LingnanCourseService
 import org.springframework.beans.factory.annotation.Autowired
+import org.springframework.cache.annotation.CacheEvict
 import org.springframework.stereotype.Controller
 import org.springframework.web.bind.annotation.*
+import java.io.File
 
 @Controller
 @RequestMapping("/api/mng")
-class ManagementController(@Autowired private val lingnanCourseService: LingnanCourseService) {
+open class ManagementController(@Autowired private val lingnanCourseService: LingnanCourseService) {
 
+    /**
+     *
+     * Get term list
+     *
+     */
     @GetMapping("/terms")
     @ResponseBody
     fun termList(): MutableMap<String, String>? = lingnanCourseService.termList
 
+    /**
+     *
+     * Clear cache and get term list
+     *
+     */
     @GetMapping("/terms", params = ["refresh"])
     @ResponseBody
     fun termListRefresh(): MutableMap<String, String>? = lingnanCourseService.reloadAndGetTermList()
 
+    /**
+     *
+     * List all tasks
+     *
+     */
     @GetMapping("/task")
     @ResponseBody
     fun taskList(): List<CaptureTaskDTO> {
         return lingnanCourseService.listTasks()
     }
 
+    /**
+     *
+     * create capture task
+     *
+     */
     @PostMapping("/task")
     @ResponseBody
     fun createTask(@RequestParam("termCode") termCode: String) =
@@ -37,11 +60,21 @@ class ManagementController(@Autowired private val lingnanCourseService: LingnanC
                 }
             }
 
+    /**
+     *
+     * Start a task
+     *
+     */
     @PostMapping("/task/{id}", params = ["start"])
     @ResponseBody
     fun startTask(@PathVariable("id") id: String): CaptureTaskDTO? =
             lingnanCourseService.startTask(id)
 
+    /**
+     *
+     * Stop a capture task
+     *
+     */
     @PostMapping("/task/{id}", params = ["stop"])
     @ResponseBody
     fun stopTask(@PathVariable("id") id: String) = HashMap<String, Any>().apply {
@@ -55,7 +88,12 @@ class ManagementController(@Autowired private val lingnanCourseService: LingnanC
         }
     }
 
-    @PostMapping("/task/{id}",params = ["import"])
+    /**
+     *
+     * Start import task data to DB
+     *
+     */
+    @PostMapping("/task/{id}", params = ["import"])
     @ResponseBody
     fun importTask(@PathVariable("id") id: String) =
             HashMap<String, Any>().apply {
@@ -92,21 +130,57 @@ class ManagementController(@Autowired private val lingnanCourseService: LingnanC
                 this["data"] = lingnanCourseService.importSubjectData(captureTaskDTO);
             }
 
+    /**
+     *
+     * Delete a task
+     *
+     */
     @DeleteMapping("/task/{id}")
     @ResponseBody
-    fun deleteTask(id : String) =
-            HashMap<String,Any>().apply {
+    fun deleteTask(@PathVariable("id") id: String) =
+            HashMap<String, Any>().apply {
+                val captureTaskResult = lingnanCourseService.queryTask(id)
+                if (captureTaskResult == null) {
+                    this["error"] = "delete_task_failed"
+                    this["message"] = "task_not_found"
+                    return this
+                }
+
+                if (captureTaskResult.folderExist) File(captureTaskResult.tempDir).deleteRecursively()
                 lingnanCourseService.deleteTask(id)
-                this["message"]="success"
+                this["message"] = "success"
             }
 
+    /**
+     *
+     * Get current running task counts
+     *
+     */
     @GetMapping("/dashes")
     @ResponseBody
     fun dashes() =
-            HashMap<String,Any>().apply {
+            HashMap<String, Any>().apply {
                 this["importingTaskCount"] = lingnanCourseService.importingTaskCount
                 this["capturingTaskCount"] = lingnanCourseService.capturingTaskCount
             }
 
-
+    /**
+     *
+     * Force clear all cache
+     *
+     */
+    @PostMapping("/cache", params = ["clear"])
+    @ResponseBody
+    @CacheEvict(
+            CacheKeys.CAPTURE_TERM_LIST,
+            CacheKeys.TERM_LIST,
+            CacheKeys.TERM_TEACHER_LIST,
+            CacheKeys.TERM_CLASS_LIST,
+            CacheKeys.TERM_COURSE_LIST,
+            CacheKeys.TERM_WEEK_RANGE,
+            allEntries = true)
+    open fun clearCache() =
+            HashMap<String, Any>().apply {
+                this["message"] = "success"
+            }
 }
